Functions are also objects in that they can
have properties and methods. A method is a property that can also be called. In other words, a method is just a function, of an object.

I had a testing ‘Ah Ha’ in that I was trying to delve into the guts of my object to test it. But I realized that I just want to get to a ‘true’ somehow. That meant I could also do something as simple as verifying a string.

“The Master plan is working just like I expected!”
- Fantastic Mr. Fox